Create HW 9 step 1

+++ src/main/java/com/anatomica/market/controllers/RestProductsController.java
@@ -75,7 +75,7 @@ public class RestProductsController {
         if (product.getId() == null || !productsService.existsById(product.getId())) {
             throw new ProductNotFoundException("Product not found, id: " + product.getId());
         }
-        if (product.getPrice() < 0) {
+        if (product.getPrice().doubleValue() < 0.0) {
             return new ResponseEntity<>("Product's price can not be negative", HttpStatus.BAD_REQUEST);
         }
         return new ResponseEntity<>(productsService.saveOrUpdate(product), HttpStatus.OK);
