+++ src/main/java/com/anatomica/market/entities/Product.java
@@ -4,6 +4,7 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Objects;
 
@@ -25,7 +26,7 @@ public class Product {
     private String description;
 
     @Column(name = "price")
-    private int price;
+    private BigDecimal price;
 
     @ManyToOne
     @JoinColumn(name = "category" )
@@ -39,7 +40,7 @@ public class Product {
 //    )
 //    private List<Category> categories;
 
-    public Product(String title, String description, int price, Category category) {
+    public Product(String title, String description, BigDecimal price, Category category) {
         this.title = title;
         this.description = description;
         this.price = price;
